Fixes #10191 - Set passed container parameters properly, BZ1204923
